Use Mastodon's doorkeeper
These are some small changes to get the plugin to work with Mastodon as
single-sign-on.

Unfortunately email addresses are not available from Mastodon, and they
are used by the oauth plugin that this plugin depends on. So it
constructs an email address using the Mastodon username and the Mastodon
hostname.

It would be nice if the email address was available from Mastodon when
doing the verify_credentials call!